refactor: MVC 패키지로 변경
