--- src/main/kotlin/lotto/Lotto.kt
@@ -1,5 +1,0 @@
-package lotto
-
-class Lotto {
-    val numbers: List<Int> = (1..45).shuffled().take(6)
-}
