--- src/main/kotlin/lotto/LottoMachine.kt
@@ -1,13 +1,0 @@
-package lotto
-
-class LottoMachine {
-    val price: Int = LOTTO_PRICE
-
-    fun getLottoCount(purchaseAmount: Int): Int {
-        return purchaseAmount / LOTTO_PRICE
-    }
-
-    companion object {
-        const val LOTTO_PRICE: Int = 1_000
-    }
-}
