+++ src/test/kotlin/lotto/domain/LottoTest.kt
@@ -1,6 +1,5 @@
 package lotto.domain
 
-import lotto.Lotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
